requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.1
python-dotenv==1.2.1
python-jose==3.5.0
passlib==1.7.4
sentence-transformers==5.1.2
torch==2.2.2
transformers==4.57.1
numpy==1.26.4
chromadb==1.3.4
huggingface_hub==0.36.0
langchain-text-splitters==1.0.0
PyPDF2==3.0.1
pydantic==2.12.4
httpx==0.28.1
orjson==3.11.4
